Minor use right type for volume texture clear value.
This has no effect in binary.

+++ src/graphics/volume_manager.cpp
@@ -39,7 +39,7 @@ void VolumeManager::updateStorage(Gl *gl)
   gl->glTexImage3D(GL_TEXTURE_3D, 0, GL_R32F, volumeAtlasSize.x(),
                    volumeAtlasSize.y(), volumeAtlasSize.z(), 0, GL_RED,
                    GL_FLOAT, nullptr);
-  int zero = 0;
+  float zero = 0.0f;
   gl->glClearTexImage(texture, 0, GL_RED, GL_FLOAT, &zero);
 
   int id = 1;
